Uploading Retrospect uninstaller script
Something I liked about the old Installer VISE Retrospect backup client
installer was that it had an “Uninstall” function. That ability was lost
when Retrospect’s builders switched to using standard Apple .pkg
installers. There was an unofficial AppleScript uninstaller that was
posted in the Retrospect forums, but it didn’t address the issue of
10.6′s receipts being stored in a different directory. So I’ve built
this script to take care of uninstalling Retrospect 6.x clients.