Updating the Interactive AD Home Folder lookup script so that it works on 10.7.x AD-bound Macs
The Interactive AD Home Folder lookup script needed to be modified for
AD lookups in 10.7, as the dscl lookup had changed from "/Active\
Directory/All\ Domains/Users/" to "/Active\ Directory/DOMAIN/All\
Domains/Users". The DOMAIN value is a stand-in for the specific AD
domain name needed by the lookup.