down to only one if statement
By checking for both conditions using AND we can simplify so that 7 & 8
pass the test. Checking for 9 and that extra code made me lol but it is
ultimately clutter. The output of spctl at this time is so simple we
only need to get the second field. This is also clear enough that we
don't need to assign result with synonymous words