Posting script and launchdaemon that sets XProtect.meta.plist to use current Java version.
Posting script and launchdaemon that sets XProtect.meta.plist to use
current Java version. This script will check the current Java 6 and
Java 7 browser plug-in versions and compare them against the minimum
version allowed by Apple's XProtect malware protection.
+
+ If the minimum Java version allowed by XProtect does not allow the
current version of the Java browser plug-in on the Mac, the script will
alter the Mac's
/System/Library/CoreServices/CoreTypes.bundle/Contents/Resources/XProtec
t.meta.plist file to set the minimum version allowed to match the
current version of the Mac's Java browser plug-in. This change allows
the Mac's current Java browser plug-in to run in Safari without being
blocked.