Updating the enable_web_java_plugin script for Safari 6.0.3 and later.
One of the features of Apple’s Safari 6.0.3 update is that it turns off
the automatic execution of Java applets through Safari, even if the
Java browser plug-in is otherwise enabled.

Safari 6.0.3 does allow for the automatic execution of Java applets to
be re-enabled through the browser. However, if it’s been a while since
a Java applet was launched, then automatic execution of Java applets is
once again automatically disabled.

I've updated my existing enable_web_java_plugin script to also handle
re-enabling Java browser plug-ins in Safari.