change to bash and use [[ vs [ - see extended
If someone has used Greg Neagle's suggestion and deleted the JavaWebComponentVersionMinimum entry from XProtect.meta.plist the above code will result in the errors when XPROTECT_JAVA_6_BUILD is empty.

I have documented the behavior in this commit to my own tweak of the above code. It was the one instance of [ that I overlooked. :(

https://github.com/rmanly/xprotect_flash_java_fixer/commit/56a5f8e7d18994ab1340246594f0df965e6b35e9