Adding safari_java_whitelist_firstboot and safari_set_java_whitelist_at_login scripts
Safari 6.0.4 and later (for Mac OS X 10.7.x and 10.8.x), and 5.1.9 and
later (for Mac OS X 10.6.x) now prompts you to enable the Java browser
plug-in on a website-by-website basis. When a Java applet is allowed,
it is added to a whitelist in Safari.

safari_java_whitelist_firstboot - This script is designed to be a
firstboot script. It sets the Safari Java whitelist settings in your
Mac's default user template and for all existing users. Currently, it
will add two servers to the Safari Java whitelist settings.

safari_set_java_whitelist_at_login - The script will add two servers to
the Safari Java whitelist settings. If the servers are already in the
whitelist, it will note that in the log, then exit.