Removed "if [ "$ENCRYPTION" = "AES-XTS" ]; then" loop from the 10.8-specific check in the FileVault encryption check scripts.
Removed the "if [ "$ENCRYPTION" = "AES-XTS" ]; then" loop from the
10.8-specific check in the FileVault encryption check scripts. Change
affects the standalone FileVault encryption check script, the Casper
Extension Attribute and the Absolute Manage Custom Info Item.

The reason for the change is that when the 10.8 CoreStorage volume
being checked is decrypting, diskutil cs list will list "Encryption
Type: None" instead of "Encryption Type: AES-XTS". As a result, no
results were being returned from the script when the machine was
decrypting.