I've updated the create_vmware_osx_install_dmg.sh script that I had previously posted.
I've updated the create_vmware_osx_install_dmg.sh script that I had
previously posted about. It now includes the ability to create an .iso
file that can be used with VMware ESXi servers running on Apple
hardware. See below the jump for the details.

When running the script, the user will now be asked if they want an ISO
disk image for use with VMware ESXi.

If the user selects Yes, both a .dmg and an .iso file will be created
and stored in the output directory.

If the user selects No, the message "-- ISO disk image will not be
created. Proceeding.." will be displayed and a .dmg file will be
created and stored in the output directory.

The virtual drives created by VMware ESXi for OS X VMs are not
automatically formatted, so support for System Image Utilty's
AutoPartition.app has been included. AutoPartition.app runs before the
installation of OS X to make sure the boot drive is formatted and named
correctly for the automated installation.