Posted updated README for iCloud pop-up disabling script
Starting in 10.7.2, Apple set the iCloud sign-in to pop up on the first
login. Since having this appear may not be desirable in all Mac
environments, it makes sense to be able to turn this off for new user
accounts.

Apple is using
/Users/username/Library/Preferences/com.apple.SetupAssistant.plist to
store  the settings that indicate whether or not the iCloud sign-in
process has run. Building on work done by the folks behind
DeployStudio, I've built a script that pre-sets those values for new
and existing accounts on a particular Mac. In turn, that should stop
the iCloud pop-up message from appearing on that Mac.