Added check to see if NVRAM entry for SIP exists
Previously you were checking only for `csrutil status`. However it is possible now that SIP has been reset (re-enabled) using the `csrutil clear` command, but that the computer has not been restarted. In that situation, the status will still show SIP as disabled, but if you check NVRAM you will see that there is no entry for "csr-active-config". This would indicate that SIP is set to be enabled after a restart. This is relatively new with 10.12.2.

https://onemoreadmin.wordpress.com/2016/12/13/system-integrity-protection-sip-changes-in-macos-sierra-10-12-2/