Uploading updated MigrateADMobileAccounttoLocalAccount migration script
Uploading updated MigrateADMobileAccounttoLocalAccount migration
script. Version is now bumped to 1.1 and includes the following changes:

1. After conversion, the specified account is added to the staff group.

All local accounts on this Mac are members of the staff group, but AD
mobile accounts are not members of the staff group.

2. The "accounttype" variable is now checking the
AuthenticationAuthority attribute instead of the OriginalNodeName
attribute. The reason for Change 2's attributes change is that the
AuthenticationAuthority attribute will exist following the conversion
process while the OriginalNodeName attribute may not.